pyproject.toml
[tool.poetry]
name = "bookstore-api"
version = "0.1.0"
description = "Classic bookstore web api example"
authors = ["MT"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
pydantic = {extras = ["all"], version = "^2.7.1"}
pydantic-settings = "^2.2.1"
sqlmodel = {extras = ["all"], version = "^0.0.18"}
asyncpg = {extras = ["all"], version = "^0.29.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
orjson = "^3.10.3"
email-validator = "^2.1.1"
alembic = {extras = ["all"], version = "^1.13.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.3"
python-logging-loki = "^0.3.1"
prometheus-client = "^0.20.0"



[tool.poetry.group.tests.dependencies]
pytest = {extras = ["all"], version = "^8.2.0"}
black = {extras = ["d"], version = "^24.4.2"}


[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["all"], version = "^0.29.0"}

[tool.poetry.group.test.dependencies]
pytest = {extras = ["all"], version = "^8.2.0"}
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
